--- CRUD-BEDROCK/Grupo2/Grupo2/src/main/java/com/example/grupo2/Beans/DetallesIncidencia.java
@@ -1,96 +1,0 @@
-package com.example.grupo2.Beans;
-
-public class DetallesIncidencia {
-
-    private  int idIncidencia;
-    private String nombreIncidencia;
-    private String descripcion;
-    private String lugar;
-    private String referencia;
-    private String tipoIncidencia;
-    private String contacto;
-    private int requiereAmbulancia;
-    private String usuario;
-    private byte[] foto;
-
-
-    public int getIdIncidencia() {
-        return idIncidencia;
-    }
-
-    public void setIdIncidencia(int idIncidencia) {
-        this.idIncidencia = idIncidencia;
-    }
-
-    public String getNombreIncidencia() {
-        return nombreIncidencia;
-    }
-
-    public void setNombreIncidencia(String nombreIncidencia) {
-        this.nombreIncidencia = nombreIncidencia;
-    }
-
-    public String getTipoIncidencia() {
-        return tipoIncidencia;
-    }
-
-    public void setTipoIncidencia(String tipoIncidencia) {
-        this.tipoIncidencia = tipoIncidencia;
-    }
-
-    public String getDescripcion() {
-        return descripcion;
-    }
-
-    public void setDescripcion(String descripcion) {
-        this.descripcion = descripcion;
-    }
-
-    public String getReferencia() {
-        return referencia;
-    }
-
-    public void setReferencia(String referencia) {
-        this.referencia = referencia;
-    }
-
-    public String getLugar() {
-        return lugar;
-    }
-
-    public void setLugar(String lugar) {
-        this.lugar = lugar;
-    }
-
-    public String getContacto() {
-        return contacto;
-    }
-
-    public void setContacto(String contacto) {
-        this.contacto = contacto;
-    }
-
-    public int getRequiereAmbulancia() {
-        return requiereAmbulancia;
-    }
-
-    public void setRequiereAmbulancia(int requiereAmbulancia) {
-        this.requiereAmbulancia = requiereAmbulancia;
-    }
-
-    public byte[] getFoto() {
-        return foto;
-    }
-
-    public void setFoto(byte[] foto) {
-        this.foto = foto;
-    }
-
-    public String getUsuario() {
-        return usuario;
-    }
-
-    public void setUsuario(String usuario) {
-        this.usuario = usuario;
-    }
-}
