--- CRUD-BEDROCK/Grupo2/Grupo2/src/main/java/com/example/grupo2/Beans/NombreUsuario.java
@@ -1,24 +1,0 @@
-package com.example.grupo2.Beans;
-
-public class NombreUsuario {
-
-    private int idUsuario;
-    private String nombreUsuario;
-
-
-    public int getIdUsuario() {
-        return idUsuario;
-    }
-
-    public void setIdUsuario(int idUsuario) {
-        this.idUsuario = idUsuario;
-    }
-
-    public String getNombreUsuario() {
-        return nombreUsuario;
-    }
-
-    public void setNombreUsuario(String nombreUsuario) {
-        this.nombreUsuario = nombreUsuario;
-    }
-}
